Move len-checker helpers to utils file
