+++ pylint/checkers/refactoring/len_checker.py
@@ -2,34 +2,10 @@
 
 # Licensed under the GPL: https://www.gnu.org/licenses/old-licenses/gpl-2.0.html
 # For details: https://github.com/PyCQA/pylint/blob/master/COPYING
-from typing import Optional
-
 import astroid
 
 from pylint import checkers, interfaces
 from pylint.checkers import utils
-
-
-def _is_call_of_name(node: astroid.node_classes.NodeNG, name: str) -> bool:
-    """Checks if node is a function call with the given name"""
-    return (
-        isinstance(node, astroid.Call)
-        and isinstance(node.func, astroid.Name)
-        and node.func.name == name
-    )
-
-
-def _is_test_condition(
-    node: astroid.node_classes.NodeNG,
-    parent: Optional[astroid.node_classes.NodeNG] = None,
-) -> bool:
-    """Returns true if the given node is being tested for truthiness"""
-    parent = parent or node.parent
-    if isinstance(parent, (astroid.While, astroid.If, astroid.IfExp, astroid.Assert)):
-        return node is parent.test or parent.test.parent_of(node)
-    if isinstance(parent, astroid.Comprehension):
-        return node in parent.ifs
-    return _is_call_of_name(parent, "bool") and parent.parent_of(node)
 
 
 class LenChecker(checkers.BaseChecker):
@@ -79,7 +55,7 @@ class LenChecker(checkers.BaseChecker):
         # a len(S) call is used inside a test condition
         # could be if, while, assert or if expression statement
         # e.g. `if len(S):`
-        if _is_call_of_name(node, "len"):
+        if utils.is_call_of_name(node, "len"):
             # the len() call could also be nested together with other
             # boolean operations, e.g. `if z or len(x):`
             parent = node.parent
@@ -88,7 +64,7 @@ class LenChecker(checkers.BaseChecker):
 
             # we're finally out of any nested boolean operations so check if
             # this len() call is part of a test condition
-            if _is_test_condition(node, parent):
+            if utils.is_test_condition(node, parent):
                 self.add_message("len-as-condition", node=node)
 
     @utils.check_messages("len-as-condition")
@@ -99,6 +75,6 @@ class LenChecker(checkers.BaseChecker):
         if (
             isinstance(node, astroid.UnaryOp)
             and node.op == "not"
-            and _is_call_of_name(node.operand, "len")
+            and utils.is_call_of_name(node.operand, "len")
         ):
             self.add_message("len-as-condition", node=node)
